Tidy up more integral types.
For #65.

Reviewed-on: https://code-review.googlesource.com/3811

+++ re2/filtered_re2.h
@@ -76,7 +76,7 @@ class FilteredRE2 {
                      vector<int>* potential_regexps) const;
 
   // The number of regexps added.
-  int NumRegexps() const { return re2_vec_.size(); }
+  int NumRegexps() const { return static_cast<int>(re2_vec_.size()); }
 
  private:
