+++ re2/onepass.cc
@@ -331,7 +331,8 @@ done:
   if (!matched)
     return false;
   for (int i = 0; i < nmatch; i++)
-    match[i].set(matchcap[2*i], matchcap[2*i+1] - matchcap[2*i]);
+    match[i].set(matchcap[2*i],
+                 static_cast<int>(matchcap[2*i+1] - matchcap[2*i]));
   return true;
 }
